requirements.txt
Flask==3.0.0
boto3==1.34.0
python-dotenv==1.0.0
Werkzeug==3.0.1
